requirements.txt
boto==2.45.0
Django==1.10.4
django-bootstrap3==7.1.0
django-storages==1.5.1
Pillow==3.4.2
six==1.10.0
